+++ qiskit_aqt_provider/circuit_to_aqt.py
@@ -14,11 +14,12 @@ from typing import List
 
 from numpy import pi
 from qiskit import QuantumCircuit
+from typing_extensions import assert_never
 
-from qiskit_aqt_provider import api_models
+from qiskit_aqt_provider import api_models, api_models_generated
 
 
-def _qiskit_to_aqt_circuit(circuit: QuantumCircuit) -> api_models.Circuit:
+def qiskit_to_aqt_circuit(circuit: QuantumCircuit) -> api_models.Circuit:
     """Convert a Qiskit `QuantumCircuit` into a payload for AQT's quantum_circuit job type.
 
     Args:
@@ -27,58 +28,91 @@ def _qiskit_to_aqt_circuit(circuit: QuantumCircuit) -> api_models.Circuit:
     Returns:
         AQT API circuit payload.
     """
-    count = 0
-    qubit_map = {}
-    for bit in circuit.qubits:
-        qubit_map[bit] = count
-        count += 1
     ops: List[api_models.OperationModel] = []
-
     num_measurements = 0
 
     for instruction in circuit.data:
-        inst = instruction[0]
-        qubits = [qubit_map[bit] for bit in instruction[1]]
-
-        if inst.name != "measure" and num_measurements > 0:
+        if instruction.operation.name != "measure" and num_measurements > 0:
             raise ValueError(
                 "Measurement operations can only be located at the end of the circuit."
             )
 
-        if inst.name == "rz":
+        if instruction.operation.name == "rz":
+            (phi,) = instruction.operation.params
+            (qubit,) = instruction.qubits
             ops.append(
                 api_models.Operation.rz(
-                    phi=float(inst.params[0]) / pi,
-                    qubit=qubits[0],
+                    phi=float(phi) / pi,
+                    qubit=circuit.find_bit(qubit).index,
                 )
             )
-        elif inst.name == "r":
+        elif instruction.operation.name == "r":
+            theta, phi = instruction.operation.params
+            (qubit,) = instruction.qubits
             ops.append(
                 api_models.Operation.r(
-                    phi=float(inst.params[1]) / pi,
-                    theta=float(inst.params[0]) / pi,
-                    qubit=qubits[0],
+                    phi=float(phi) / pi,
+                    theta=float(theta) / pi,
+                    qubit=circuit.find_bit(qubit).index,
                 )
             )
-        elif inst.name == "rxx":
+        elif instruction.operation.name == "rxx":
+            (theta,) = instruction.operation.params
+            q0, q1 = instruction.qubits
             ops.append(
                 api_models.Operation.rxx(
-                    theta=float(inst.params[0]) / pi,
-                    qubits=qubits[:2],
+                    theta=float(theta) / pi,
+                    qubits=[circuit.find_bit(q0).index, circuit.find_bit(q1).index],
                 )
             )
-        elif inst.name == "measure":
+        elif instruction.operation.name == "measure":
             num_measurements += 1
-        elif inst.name == "barrier":
+        elif instruction.operation.name == "barrier":
             continue
         else:
-            raise ValueError(f"Operation '{inst.name}' not in basis gate set: {{rz, r, rxx}}")
+            raise ValueError(
+                f"Operation '{instruction.operation.name}' not in basis gate set: {{rz, r, rxx}}"
+            )
 
     if not num_measurements:
         raise ValueError("Circuit must have at least one measurement operation.")
 
     ops.append(api_models.Operation.measure())
     return api_models.Circuit(root=ops)
+
+
+def aqt_to_qiskit_circuit(circuit: api_models.Circuit, number_of_qubits: int) -> QuantumCircuit:
+    """Convert an AQT API quantum circuit payload to an equivalent Qiskit representation.
+
+    Args:
+        circuit: payload to convert
+        number_of_qubits: size of the quantum register to use for the converted circuit.
+
+    Returns:
+        A :class:`QuantumCircuit <qiskit.circuit.quantumcircuit.QuantumCircuit>` equivalent
+        to the passed circuit payload.
+    """
+    qiskit_circuit = QuantumCircuit(number_of_qubits)
+
+    for operation in circuit.__root__:
+        if isinstance(operation.__root__, api_models_generated.GateRZ):
+            qiskit_circuit.rz(operation.__root__.phi * pi, operation.__root__.qubit)
+        elif isinstance(operation.__root__, api_models_generated.GateR):
+            qiskit_circuit.r(
+                operation.__root__.theta * pi,
+                operation.__root__.phi * pi,
+                operation.__root__.qubit,
+            )
+        elif isinstance(operation.__root__, api_models_generated.GateRXX):
+            qiskit_circuit.rxx(
+                operation.__root__.theta * pi, *[mod.__root__ for mod in operation.__root__.qubits]
+            )
+        elif isinstance(operation.__root__, api_models_generated.Measure):
+            qiskit_circuit.measure_all()
+        else:
+            assert_never(operation.__root__)  # pragma: no cover
+
+    return qiskit_circuit
 
 
 def circuits_to_aqt_job(circuits: List[QuantumCircuit], shots: int) -> api_models.JobSubmission:
@@ -98,7 +132,7 @@ def circuits_to_aqt_job(circuits: List[QuantumCircuit], shots: int) -> api_model
             circuits=[
                 api_models.QuantumCircuit(
                     repetitions=shots,
-                    quantum_circuit=_qiskit_to_aqt_circuit(circuit),
+                    quantum_circuit=qiskit_to_aqt_circuit(circuit),
                     number_of_qubits=circuit.num_qubits,
                 )
                 for circuit in circuits