+++ qiskit_aqt_provider/test/fixtures.py
@@ -20,10 +20,12 @@ from typing import List, Tuple
 
 import pytest
 from qiskit.circuit import QuantumCircuit
+from typing_extensions import override
 
+from qiskit_aqt_provider import api_models
+from qiskit_aqt_provider.aqt_job import AQTJob
 from qiskit_aqt_provider.aqt_provider import AQTProvider
 from qiskit_aqt_provider.aqt_resource import OfflineSimulatorResource
-from qiskit_aqt_provider.circuit_to_aqt import _qiskit_to_aqt_circuit
 
 
 class MockSimulator(OfflineSimulatorResource):
@@ -32,38 +34,28 @@ class MockSimulator(OfflineSimulatorResource):
     def __init__(self, *, noisy: bool) -> None:
         super().__init__(
             AQTProvider(""),
-            workspace_id="default",
-            resource_id="mock_simulator",
-            resource_name="mock_simulator",
-            noisy=noisy,
+            resource_id=api_models.ResourceId(
+                workspace_id="default",
+                resource_id="mock_simulator",
+                resource_name="mock_simulator",
+                resource_type="offline_simulator",
+            ),
+            with_noise_model=noisy,
         )
 
         self.submit_call_args: List[Tuple[List[QuantumCircuit], int]] = []
 
-    def submit(self, circuits: List[QuantumCircuit], shots: int) -> uuid.UUID:
+    @override
+    def submit(self, job: AQTJob) -> uuid.UUID:
         """Submit the circuits for execution on the backend.
 
         Record the passed arguments in `submit_call_args`.
 
-        Try to convert the circuits to the AQT JSON wire format.
-
         Args:
-            circuits: the circuits to execute on the simulator
-            shots: number of repetitions.
-
-        Raises:
-            ValueError: at least one circuit cannot be converted to the AQT JSON wire format.
+            job: AQTJob to submit to the mock simulator.
         """
-        for circuit in circuits:
-            try:
-                _ = _qiskit_to_aqt_circuit(circuit)
-            except Exception as e:  # noqa: BLE001
-                raise ValueError(
-                    f"Circuit cannot be converted to AQT JSON format:\n{circuit}"
-                ) from e
-
-        self.submit_call_args.append((circuits, shots))
-        return super().submit(circuits, shots)
+        self.submit_call_args.append((job.circuits, job.options.shots))
+        return super().submit(job)
 
     @property
     def submitted_circuits(self) -> List[List[QuantumCircuit]]: