+++ test/test_resource.py
@@ -13,11 +13,14 @@
 import json
 import math
 import uuid
+from contextlib import nullcontext
+from typing import Any, ContextManager
 from unittest import mock
 
 import httpx
 import pydantic as pdt
 import pytest
+import qiskit
 from polyfactory.factories.pydantic_factory import ModelFactory
 from pytest_httpx import HTTPXMock
 from qiskit import QuantumCircuit
@@ -27,7 +30,8 @@ from qiskit_aqt_provider import api_models
 from qiskit_aqt_provider.aqt_job import AQTJob
 from qiskit_aqt_provider.aqt_options import AQTOptions
 from qiskit_aqt_provider.aqt_resource import AQTResource
-from qiskit_aqt_provider.test.circuits import assert_circuits_equal, empty_circuit
+from qiskit_aqt_provider.circuit_to_aqt import circuits_to_aqt_job
+from qiskit_aqt_provider.test.circuits import assert_circuits_equal, empty_circuit, random_circuit
 from qiskit_aqt_provider.test.fixtures import MockSimulator
 from qiskit_aqt_provider.test.resources import DummyResource, TestResource
 from qiskit_aqt_provider.versions import USER_AGENT
@@ -97,8 +101,9 @@ def test_query_timeout_propagation() -> None:
 
     qc = QuantumCircuit(1)
     qc.rx(3.14, 0)
+    qc.measure_all()
 
-    job = backend.run(qc)
+    job = backend.run(qiskit.transpile(qc, backend))
 
     with pytest.raises(JobTimeoutError):
         job.result()
@@ -125,7 +130,7 @@ def test_query_period_propagation() -> None:
     qc.rx(3.14, 0)
     qc.measure_all()
 
-    job = backend.run(qc)
+    job = backend.run(qiskit.transpile(qc, backend))
 
     with mock.patch.object(AQTJob, "status", wraps=job.status) as mocked_status:
         job.result()
@@ -236,16 +241,53 @@ def test_submit_valid_response(httpx_mock: HTTPXMock) -> None:
             json=json.loads(
                 api_models.Response.queued(
                     job_id=expected_job_id,
-                    resource_id=backend.resource_id,
-                    workspace_id=backend.workspace_id,
+                    resource_id=backend.resource_id.resource_id,
+                    workspace_id=backend.resource_id.workspace_id,
                 ).model_dump_json()
             ),
         )
 
     httpx_mock.add_callback(handle_submit, method="POST")
 
-    job_id = backend.submit([empty_circuit(2)], shots=10)
-    assert job_id == expected_job_id
+    job = AQTJob(backend, circuits=[empty_circuit(2)], options=AQTOptions(shots=10))
+    job.submit()
+    assert job.job_id() == str(expected_job_id)
+
+
+def test_submit_payload_matches(httpx_mock: HTTPXMock) -> None:
+    """Check that the quantum circuits jobs payload is correctly submitted to the API endpoint."""
+    backend = DummyResource("")
+    shots = 123
+    qc = qiskit.transpile(random_circuit(2), backend)
+    expected_job_payload = circuits_to_aqt_job([qc], shots=shots)
+    expected_job_id = uuid.uuid4()
+
+    def handle_submit(request: httpx.Request) -> httpx.Response:
+        assert request.headers["user-agent"] == USER_AGENT
+        assert request.url.path.endswith(
+            f"submit/{backend.resource_id.workspace_id}/{backend.resource_id.resource_id}"
+        )
+
+        data = api_models.JobSubmission.parse_raw(request.content.decode("utf-8"))
+        assert data == expected_job_payload
+
+        return httpx.Response(
+            status_code=httpx.codes.OK,
+            json=json.loads(
+                api_models.Response.queued(
+                    job_id=expected_job_id,
+                    resource_id=backend.resource_id.resource_id,
+                    workspace_id=backend.resource_id.workspace_id,
+                ).json()
+            ),
+        )
+
+    httpx_mock.add_callback(handle_submit, method="POST")
+
+    qc = qiskit.transpile(random_circuit(2), backend)
+    job = AQTJob(backend, circuits=[qc], options=AQTOptions(shots=shots))
+    job.submit()
+    assert job.job_id() == str(expected_job_id)
 
 
 def test_submit_bad_request(httpx_mock: HTTPXMock) -> None:
@@ -255,8 +297,9 @@ def test_submit_bad_request(httpx_mock: HTTPXMock) -> None:
     backend = DummyResource("")
     httpx_mock.add_response(status_code=httpx.codes.BAD_REQUEST)
 
+    job = AQTJob(backend, circuits=[empty_circuit(2)], options=AQTOptions(shots=10))
     with pytest.raises(httpx.HTTPError):
-        backend.submit([empty_circuit(2)], shots=10)
+        job.submit()
 
 
 def test_result_valid_response(httpx_mock: HTTPXMock) -> None:
@@ -268,12 +311,15 @@ def test_result_valid_response(httpx_mock: HTTPXMock) -> None:
     job_id = uuid.uuid4()
 
     payload = api_models.Response.cancelled(
-        job_id=job_id, resource_id=backend.resource_id, workspace_id=backend.workspace_id
+        job_id=job_id,
+        resource_id=backend.resource_id.resource_id,
+        workspace_id=backend.resource_id.workspace_id,
     )
 
     def handle_result(request: httpx.Request) -> httpx.Response:
         assert request.headers["user-agent"] == USER_AGENT
         assert request.headers["authorization"] == f"Bearer {token}"
+        assert request.url.path.endswith(f"result/{job_id}")
 
         return httpx.Response(
             status_code=httpx.codes.OK, json=json.loads(payload.model_dump_json())
@@ -311,16 +357,126 @@ def test_result_unknown_job(httpx_mock: HTTPXMock) -> None:
         backend.result(job_id)
 
 
-def test_resource_fixture_detect_invalid_circuits(
+def test_offline_simulator_detects_invalid_circuits(
     offline_simulator_no_noise: MockSimulator,
 ) -> None:
-    """Pass a circuit that cannot be converted to the AQT API to the mock simulator.
+    """Pass a circuit that cannot be converted to the AQT API to the offline simulator.
     This must fail.
     """
     qc = QuantumCircuit(2)
     qc.h(0)
-    qc.cnot(0, 1)
+    qc.cx(0, 1)
     qc.measure_all()
 
-    with pytest.raises(ValueError, match="^Circuit cannot be converted"):
+    with pytest.raises(ValueError, match="^Operation 'h' not in basis gate set"):
         offline_simulator_no_noise.run(qc)
+
+
+def test_offline_simulator_propagate_shots_option(
+    offline_simulator_no_noise: MockSimulator,
+) -> None:
+    """Check various ways of configuring the number of repetitions."""
+    qc = qiskit.transpile(random_circuit(2), offline_simulator_no_noise)
+
+    default_shots = sum(offline_simulator_no_noise.run(qc).result().get_counts().values())
+    assert default_shots == AQTOptions().shots
+
+    # TODO: use annotated-types to get unified access to upper bound
+    shots = min(default_shots + 40, AQTOptions.__fields__["shots"].field_info.le)
+    assert shots != default_shots
+
+    # configure shots in AQTResource.run
+    shots_run = sum(offline_simulator_no_noise.run(qc, shots=shots).result().get_counts().values())
+    assert shots_run == shots
+
+    # configure shots in qiskit.execute
+    shots_execute = sum(
+        qiskit.execute(qc, offline_simulator_no_noise, shots=shots).result().get_counts().values()
+    )
+    assert shots_execute == shots
+
+    # configure shots in resource options
+    offline_simulator_no_noise.options.shots = shots
+    shots_options = sum(offline_simulator_no_noise.run(qc).result().get_counts().values())
+    assert shots_options == shots
+
+    # qiskit.execute overrides resource options
+    shots_override = min(shots + 40, AQTOptions.__fields__["shots"].field_info.le)
+    assert shots_override != shots
+    assert shots_override != default_shots
+    assert offline_simulator_no_noise.options.shots != shots_override
+    shots = sum(
+        qiskit.execute(qc, offline_simulator_no_noise, shots=shots_override)
+        .result()
+        .get_counts()
+        .values()
+    )
+    assert shots == shots_override
+
+
+@pytest.mark.parametrize(
+    ("memory", "context"),
+    [(True, nullcontext()), (False, pytest.raises(qiskit.QiskitError, match="No memory"))],
+)
+def test_offline_simulator_run_propagate_memory_option(
+    memory: bool,
+    context: ContextManager[Any],
+    offline_simulator_no_noise: MockSimulator,
+) -> None:
+    """Check that the memory option can be set on `AQTResource.run`."""
+    qc = qiskit.transpile(random_circuit(2), offline_simulator_no_noise)
+    default_shots = AQTOptions().shots
+
+    result = offline_simulator_no_noise.run(qc, memory=memory).result()
+    with context:
+        assert len(result.get_memory()) == default_shots
+
+
+@pytest.mark.parametrize(
+    ("memory", "context"),
+    [(True, nullcontext()), (False, pytest.raises(qiskit.QiskitError, match="No memory"))],
+)
+def test_offline_simulator_execute_propagate_memory_option(
+    memory: bool, context: ContextManager[Any], offline_simulator_no_noise: MockSimulator
+) -> None:
+    """Check that the memory option can be set in `qiskit.execute`."""
+    qc = random_circuit(2)
+    default_shots = AQTOptions().shots
+
+    result = qiskit.execute(qc, offline_simulator_no_noise, memory=memory).result()
+    with context:
+        assert len(result.get_memory()) == default_shots
+
+
+@pytest.mark.parametrize(
+    ("memory", "context"),
+    [(True, nullcontext()), (False, pytest.raises(qiskit.QiskitError, match="No memory"))],
+)
+def test_offline_simulator_resource_propagate_memory_option(
+    memory: bool, context: ContextManager[Any], offline_simulator_no_noise: MockSimulator
+) -> None:
+    """Check that the memory option can be set as resource option."""
+    qc = qiskit.transpile(random_circuit(2), offline_simulator_no_noise)
+    default_shots = AQTOptions().shots
+
+    offline_simulator_no_noise.options.memory = memory
+    result = offline_simulator_no_noise.run(qc).result()
+    with context:
+        assert len(result.get_memory()) == default_shots
+
+
+@pytest.mark.parametrize(
+    ("memory", "context"),
+    [(True, nullcontext()), (False, pytest.raises(qiskit.QiskitError, match="No memory"))],
+)
+def test_offline_simulator_execute_override_memory_option(
+    memory: bool, context: ContextManager[Any], offline_simulator_no_noise: MockSimulator
+) -> None:
+    """Check that setting `memory` through `qiskit.execute` overrides the resource options."""
+    qc = random_circuit(2)
+    default_shots = AQTOptions().shots
+
+    offline_simulator_no_noise.options.memory = not memory
+    result = qiskit.execute(qc, offline_simulator_no_noise, memory=memory).result()
+    with context:
+        assert len(result.get_memory()) == default_shots